+++ src/Section/AboutSection/BackgroundEffects.tsx
