+++ src/Section/AboutSection/EnergyPulse.tsx
