+++ src/Section/AboutSection/WaveLines.tsx
